requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
time-machine==2.13.0
streamlit==1.28.1
streamlit-authenticator==0.2.3
plotly==5.17.0
watchdog==3.0.0
bcrypt==4.1.2
extra-streamlit-components==0.1.60
streamlit-option-menu==0.3.6
streamlit-autorefresh==0.0.1